components/liveinstall: Switch to EROFS for live media
This switches from SquashFS to EROFS using LZMA compression
with a 1M chunksize and fragement dedpulication for inodes enabled.

This attempts to match the defaults for SquashFS in kiwi and provides
a good balance for image compression and creation time.

Note that "-Ededupe" is not used because it is not multi-threaded and
is currently extremely slow.